Use case_properties_changed helper
