+++ custom/enikshay/integrations/nikshay/repeaters.py
@@ -3,11 +3,10 @@ from django.core.urlresolvers import reverse
 from corehq.apps.repeaters.models import CaseRepeater
 from corehq.form_processor.models import CommCareCaseSQL
 from corehq.toggles import NIKSHAY_INTEGRATION
-from casexml.apps.case.xml.parser import CaseUpdateAction
-from casexml.apps.case.xform import get_case_updates
 from casexml.apps.case.models import CommCareCase
 from casexml.apps.case.signals import case_post_save
 from corehq.apps.repeaters.signals import create_repeat_records
+from custom.enikshay.case_utils import case_properties_changed
 
 
 class NikshayRegisterPatientRepeater(CaseRepeater):
@@ -34,23 +33,11 @@ class NikshayRegisterPatientRepeater(CaseRepeater):
         return allowed_case_types_and_users and (
             not episode_case_properties.get('nikshay_registered', 'false') == 'true' and
             not episode_case_properties.get('nikshay_id', False) and
-            episode_pending_registration_changed(episode_case)
+
+            # Episode pending registration flips from yes to no
+            case_properties_changed(episode_case, ['episode_pending_registration']) and
+            episode_case_properties.get('episode_pending_registration', 'yes') == 'no'
         )
-
-
-def episode_pending_registration_changed(case):
-    last_case_action = case.actions[-1]
-    if last_case_action.is_case_create:
-        return False
-
-    last_update_actions = [update.get_update_action() for update in get_case_updates(last_case_action.form)]
-    value_changed = any(
-        action for action in last_update_actions
-        if isinstance(action, CaseUpdateAction)
-        and 'episode_pending_registration' in action.dynamic_properties
-        and action.dynamic_properties['episode_pending_registration'] == 'no'
-    )
-    return value_changed
 
 
 def create_case_repeat_records(sender, case, **kwargs):
