+++ custom/enikshay/integrations/ninetyninedots/repeaters.py
@@ -4,14 +4,13 @@ from django.utils.translation import ugettext_lazy as _
 
 from corehq.form_processor.models import CommCareCaseSQL
 from casexml.apps.case.models import CommCareCase
-from casexml.apps.case.xml.parser import CaseUpdateAction
 
 from corehq.apps.repeaters.models import CaseRepeater
 from corehq.apps.repeaters.signals import create_repeat_records
 from casexml.apps.case.signals import case_post_save
-from casexml.apps.case.xform import get_case_updates
 
 from custom.enikshay.case_utils import (
+    case_properties_changed,
     get_open_episode_case_from_person,
     get_episode_case_from_adherence,
     CASE_TYPE_EPISODE,
@@ -21,10 +20,6 @@ from custom.enikshay.const import (
     TREATMENT_OUTCOME,
     NINETYNINEDOTS_EPISODE_PROPERTIES,
     NINETYNINEDOTS_PERSON_PROPERTIES,
-)
-from custom.enikshay.case_utils import (
-    get_occurrence_case_from_episode,
-    get_person_case_from_occurrence,
 )
 from custom.enikshay.exceptions import ENikshayCaseNotFound
 
@@ -180,25 +175,6 @@ class NinetyNineDotsTreatmentOutcomeRepeater(Base99DOTSRepeater):
 
 def episode_registered_with_99dots(episode):
     return episode.dynamic_case_properties().get('dots_99_registered', False) == 'true'
-
-
-def case_properties_changed(case, case_properties):
-    if isinstance(case_properties, basestring):
-        case_properties = [case_properties]
-
-    last_case_action = case.actions[-1]
-    if last_case_action.is_case_create:
-        return False
-
-    update_actions = [update.get_update_action() for update in get_case_updates(last_case_action.form)]
-    property_changed = any(
-        action for action in update_actions
-        if isinstance(action, CaseUpdateAction)
-        and any(
-            case_property in action.dynamic_properties for case_property in case_properties
-        )
-    )
-    return property_changed
 
 
 def create_case_repeat_records(sender, case, **kwargs):
